code for 1st part of auto paths
