+++ .idea/gradle.xml
@@ -7,6 +7,14 @@
         <option name="testRunner" value="GRADLE" />
         <option name="distributionType" value="DEFAULT_WRAPPED" />
         <option name="externalProjectPath" value="$PROJECT_DIR$" />
+        <option name="modules">
+          <set>
+            <option value="$PROJECT_DIR$" />
+            <option value="$PROJECT_DIR$/FtcRobotController" />
+            <option value="$PROJECT_DIR$/MeepMeepPathing" />
+            <option value="$PROJECT_DIR$/TeamCode" />
+          </set>
+        </option>
       </GradleProjectSettings>
     </option>
   </component>
