+++ MeepMeepPathing/src/main/java/com/example/meepmeeppathing/BlueLeft/BlueLeftAutoExact.java
@@ -1,0 +1,46 @@
+package com.example.meepmeeppathing.BlueLeft;
+
+import com.acmerobotics.roadrunner.geometry.Pose2d;
+import com.acmerobotics.roadrunner.geometry.Vector2d;
+import com.noahbres.meepmeep.MeepMeep;
+import com.noahbres.meepmeep.roadrunner.DefaultBotBuilder;
+import com.noahbres.meepmeep.roadrunner.entity.RoadRunnerBotEntity;
+
+public class BlueLeftAutoExact {
+    public static void main(String[] args) {
+        // Declare a MeepMeep instance
+        // With a field size of 800 pixels
+        MeepMeep meepMeep = new MeepMeep(800);
+
+        RoadRunnerBotEntity myBot = new DefaultBotBuilder(meepMeep)
+                // Required: Set bot constraints: maxVel, maxAccel, maxAngVel, maxAngAccel, track width
+                .setConstraints(60, 60, Math.toRadians(180), Math.toRadians(180), 13)
+                // Option: Set theme. Default = ColorSchemeRedDark()
+                .followTrajectorySequence(drive ->
+                                // place pixel 1
+                                drive.trajectorySequenceBuilder(new Pose2d(11.28, 60.36, Math.toRadians(270.00)))
+                                        .splineTo(new Vector2d(23.21, 34.56), Math.toRadians(-65.16))
+                                        .build()
+
+/*                                // place pixel 2
+                                drive.trajectorySequenceBuilder(new Pose2d(10.83, 59.03, Math.toRadians(270.00)))
+                                        .splineTo(new Vector2d(17.47, 24.25), Math.toRadians(-79.20))
+                                        .build()
+
+                                // place pixel 3
+                                drive.trajectorySequenceBuilder(new Pose2d(11.28, 59.32, Math.toRadians(270.00)))
+                                        .splineTo(new Vector2d(0.96, 34.27), Math.toRadians(247.62))
+                                        .build()*/
+
+//
+                );
+
+        // Set field image
+        meepMeep.setBackground(MeepMeep.Background.FIELD_CENTERSTAGE_OFFICIAL)
+                .setDarkMode(true)
+                // Background opacity from 0-1
+                .setBackgroundAlpha(0.95f)
+                .addEntity(myBot)
+                .start();
+    }
+}
